Separated token generator and verifier
