+++ src/auth/token-verifier.ts
@@ -26,6 +26,33 @@ import https = require('https');
 // Audience to use for Firebase Auth Custom tokens
 const FIREBASE_AUDIENCE = 'https://identitytoolkit.googleapis.com/google.identity.identitytoolkit.v1.IdentityToolkit';
 
+export const ALGORITHM_RS256 = 'RS256';
+
+// URL containing the public keys for the Google certs (whose private keys are used to sign Firebase
+// Auth ID tokens)
+const CLIENT_CERT_URL = 'https://www.googleapis.com/robot/v1/metadata/x509/securetoken@system.gserviceaccount.com';
+
+// URL containing the public keys for Firebase session cookies. This will be updated to a different URL soon.
+const SESSION_COOKIE_CERT_URL = 'https://www.googleapis.com/identitytoolkit/v3/relyingparty/publicKeys';
+
+/** User facing token information related to the Firebase ID token. */
+export const ID_TOKEN_INFO: FirebaseTokenInfo = {
+  url: 'https://firebase.google.com/docs/auth/admin/verify-id-tokens',
+  verifyApiName: 'verifyIdToken()',
+  jwtName: 'Firebase ID token',
+  shortName: 'ID token',
+  expiredErrorCode: 'auth/id-token-expired',
+};
+
+/** User facing token information related to the Firebase session cookie. */
+export const SESSION_COOKIE_INFO: FirebaseTokenInfo = {
+  url: 'https://firebase.google.com/docs/auth/admin/manage-cookies',
+  verifyApiName: 'verifySessionCookie()',
+  jwtName: 'Firebase session cookie',
+  shortName: 'session cookie',
+  expiredErrorCode: 'auth/session-cookie-expired',
+};
+
 /** Interface that defines token related user facing information. */
 export interface FirebaseTokenInfo {
   /** Documentation URL. */
@@ -46,7 +73,7 @@ export interface FirebaseTokenInfo {
 export class FirebaseTokenVerifier {
   private publicKeys: object;
   private publicKeysExpireAt: number;
-  private shortNameArticle: string;
+  private readonly shortNameArticle: string;
 
   constructor(private clientCertUrl: string, private algorithm: string,
               private issuer: string, private projectId: string,
@@ -283,4 +310,34 @@ export class FirebaseTokenVerifier {
       }).on('error', reject);
     });
   }
+}
+
+/**
+ * Creates a new FirebaseTokenVerifier to verify Firebase ID tokens.
+ *
+ * @param projectId Project ID string.
+ */
+export function newIdTokenVerifier(projectId: string): FirebaseTokenVerifier {
+  return new FirebaseTokenVerifier(
+      CLIENT_CERT_URL,
+      ALGORITHM_RS256,
+      'https://securetoken.google.com/',
+      projectId,
+      ID_TOKEN_INFO,
+  );
+}
+
+/**
+ * Creates a new FirebaseTokenVerifier to verify Firebase session cookies.
+ *
+ * @param projectId Project ID string.
+ */
+export function newSessionCookieVerifier(projectId: string): FirebaseTokenVerifier {
+  return new FirebaseTokenVerifier(
+    SESSION_COOKIE_CERT_URL,
+    ALGORITHM_RS256,
+    'https://session.firebase.google.com/',
+    projectId,
+    SESSION_COOKIE_INFO,
+  );
 }
